test_muliple dbs moved to one file. Using new verification function
